+++ tests/Neo.L2.Abstractions.UnitTests/UT_NeoHubDeployReport.cs
@@ -209,6 +209,8 @@ public class UT_NeoHubDeployReport
                 + "RpcChainId / ProofType / SettlementConfiguredProofType / IsChainIdConfigConsistent / "
                 + "IsProofTypeConfigConsistent / DaMode / RpcDaMode / IsDaModeConfigConsistent / "
                 + "IsNeoHubHashWiringComplete / IsBatcherInboxWiringComplete / "
+                + "IsSecurityLevelProofTypeConsistent / HasExpectedNetwork / HasScannerDeployHeights / "
+                + "IsOfflinePassportComplete / "
                 + "HasSealedBatchSink / NextExpectedBlock / ProcessCommittedBlock / IsOperatorReady / "
                 + "PeekSharedBridgeDeposits / GetOperatorStatusAsync",
                 stores.GetProperty("localHostReadiness").GetString());